tools.find.mfind_bin: fix dtype inconsistency and mex out-of-bounds read
mfind_bin.m:65 returned untyped double for the numel(X)<2 &&
~isempty(X) case, unlike every other branch (uint32) - fixed to
uint32(ones(1,Nvals)).

More seriously: both mex_mfind_bin_double.cpp and
mex_mfind_bin_single.cpp initialize Im to floor(0.5*(1+N)); for N==0
(empty X) that's 0, and the search loop then reads X[Im[i]-1]. Since
Im[i] is uint32_t, Im[i]-1 when Im[i]==0 underflows to UINT32_MAX, not
simply -1 - an out-of-bounds read at a huge offset past the (empty)
array, a real memory-safety bug independent of whether it's exercised
today. Added an explicit N==0 guard in both files, returning an empty
1x0 uint32 result before ever entering the search loop, matching
mfind_bin.m's own isempty(X) branch.

Verified: 1-element X now returns uint32 in the matlab path; empty X
returns an empty uint32 cleanly in both the double and single mex
kernels (previously an OOB read); normal multi-element case unaffected
(mex and matlab still agree exactly); 1-element X mex path unaffected
(was already correct - only N==0 was broken).

Fixes #1210

+++ matlab/util/+tools/+find/mfind_bin.m
@@ -62,7 +62,7 @@ else
         if (isempty(X))
             Im = uint32([]);
         else
-            Im = ones(1,Nvals);
+            Im = uint32(ones(1,Nvals));
         end
     else
         %Niter = 0;
